move PersonalDataController and related logic from project-base to framework package
