+++ packages/framework/src/Resources/views/Front/Content/PersonalData/export.xml.twig
@@ -12,7 +12,7 @@
             </customer_name>
 
             {% if customerUser.customer.billingAdress is defined or customerUser.customer.deliveryAddresses is defined %}
-                {% include 'Front/Content/PersonalData/adress.xml.twig' with {'customerUser' : customerUser} %}
+                {% include '@ShopsysFramework/Front/Content/PersonalData/adress.xml.twig' with {'customerUser' : customerUser} %}
             {% endif %}
 
             {% if customerUser.customer.billingAddress is defined and customerUser.customer.billingAddress.companyCustomer %}
@@ -36,7 +36,7 @@
         </other_information>
 
         {% if orders|length > 0 %}
-            {% include('Front/Content/PersonalData/orders.xml.twig' ) with {'orders': orders} %}
+            {% include('@ShopsysFramework/Front/Content/PersonalData/orders.xml.twig' ) with {'orders': orders} %}
         {% endif %}
 
     </customer>
